Separate methods into different classes based on functionality
Everything is in one class (except tasks which was done before)

Divide methods into classes based on their usage (i.e Storage
methods goes to Storage class, etc)

Code is more modular and clean this way
