+++ src/main/java/Lamball.java
@@ -1,336 +1,53 @@
-import java.io.File;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.time.format.DateTimeParseException;
-import java.util.List;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Scanner;
+
 public class Lamball {
-    String indent = "    ____________________________________________________________\n";
-    String welcomeMessage = "    Lamball\n" +
-            "    ____________________________________________________________\n" +
-            "     Hello! I'm Lamball, your helpful sheep!\n" +
-            "     Whaaat can I do for you?\n" +
-            "    ____________________________________________________________\n";
-    String goodbyeMessage = "    ____________________________________________________________\n" +
-            "     See you again!\n" +
-            "    ____________________________________________________________\n";
+    public static String indent = "    ____________________________________________________________\n";
 
-    ArrayList<Task> tasks;
-    File savedList;
+    private TaskList tasks;
+    private Ui ui;
+
     public Lamball() {
-        tasks = new ArrayList<>();
+        tasks = new TaskList();
+        ui = new Ui();
     }
 
-    public void greetingMessage() {
-        System.out.println(this.welcomeMessage);
-    }
-    public void goodbyeMessage() {
-        System.out.println(this.goodbyeMessage);
+    public void initialize() {
+        ui.greetingMessage();
+        Storage.obtainSavedFile(this);
     }
 
-    public void printList() {
-        System.out.println(indent);
-        System.out.println("    Here aaaaare the taaaasks in your list:");
-        for(int i = 0; i < tasks.size(); i++) {
-            System.out.println("    " + (i + 1) + ". " + tasks.get(i).toString() + "");
-        }
-        System.out.println(indent);
+    public boolean parse(String msg, boolean isInit) throws LamballParseException{
+        boolean active = Parser.parse(msg, isInit, tasks);
+        return active;
     }
 
-    private void mark(String[] parts, boolean isInit) throws LamballParseException {
-        // If not numeric
-        if (parts.length < 2 || !parts[1].matches("-?\\d+")) {
-            throw new LamballParseException(indent + "    Invalid number, baa.\n" + indent);
-        }
-        int idx = Integer.valueOf(parts[1]) - 1;
-        // Checks if index is within range of list
-        if (idx >= tasks.size() || idx < 0) {
-            throw new LamballParseException(indent + "    Taaask index out of range, baa.\n" + indent);
-        }
-        Task temp = tasks.get(idx);
-        temp.mark();
-        replaceLine("1 | " + temp.command(), idx);
-        if (!isInit) {
-            System.out.println(indent + "    I have maaarked the task as done:\n");
-            System.out.println("        " + temp.toString() + "\n" + indent);
-        }
-    }
+    public void run() {
+        Scanner scanner = new Scanner(System.in);
+        Lamball lamball = new Lamball();
+        boolean active = true;
 
-    private void unMark(String[] parts) throws LamballParseException {
-        // If not numeric
-        if (parts.length < 2 || !parts[1].matches("-?\\d+")) {
-            throw new LamballParseException(indent + "    Invalid number, baa.\n" + indent);
-        }
-        int idx = Integer.valueOf(parts[1]) - 1;
-        // Checks if index is within range of list
-        if (idx >= tasks.size() || idx < 0) {
-            throw new LamballParseException(indent + "    Taaask index out of range, baa.\n" + indent);
-        }
-        System.out.println(indent + "    I have maaarked the task as undone:\n" + indent);
-        Task temp = tasks.get(idx);
-        temp.unMark();
-        replaceLine("0 | " + temp.command(), idx);
-        System.out.println("        " + temp.toString() + "\n" + indent);
-    }
+        lamball.initialize();
 
-    private void toDo(String[] parts, boolean isInit) throws LamballParseException {
-        // Checks if empty string (nothing after command) or only whitespaces
-        if (parts.length < 2 || parts[1] == null || parts[1].trim().isEmpty()) {
-            throw new LamballParseException(indent + "    Your todo field is empty, baaaaka.\n" + indent);
-        }
-        Task temp = new ToDo(parts[1]);
-        tasks.add(temp);
-        if (!isInit) {
-            System.out.println(indent + "    Added ToDo:\n        " + temp.toString() + "\n    Now you have " + tasks.size()
-                    + " tasks in the list.\n" + indent);
-            writeToFile("0 | " + temp.command());
-        }
-    }
-    private void deadline(String[] parts, boolean isInit) throws LamballParseException {
-        // Checks if empty string (nothing after command) or only whitespaces
-        if (parts.length < 2 || parts[1] == null || parts[1].trim().isEmpty()) {
-            throw new LamballParseException(indent + "    Your deadline field is empty, baaaaka.\n" + indent);
-        }
-        String[] furtherSplit = parts[1].split(" /", 2);
-        if (furtherSplit.length < 2 || !furtherSplit[1].substring(0,3).equals("by ")) {
-            throw new LamballParseException(indent + "    Deadline is in the wrong formaaaaaaat, baa. :(\n    Correct fo" +
-                    "rmaaat is: deadline <name> /by <time>, baa.\n" + indent);
-        }
-        try {
-            Task temp = new Deadline(furtherSplit[0], furtherSplit[1].replaceFirst("by ", ""));
-            tasks.add(temp);
-            if (!isInit) {
-                System.out.println(indent + "    Added Deadline:\n        " + temp.toString() + "\n    Now you have " +
-                        tasks.size() + " tasks in the list.\n" + indent);
-                writeToFile("0 | " + temp.command());
-            }
-        } catch (DateTimeParseException e) {
-            throw new LamballParseException(indent + "    Date is in the wrong formaaaaaaat, baa. :(\n    Correct fo" +
-                    "rmaaat is: yyyy-mm-dd (e.g 2001-01-20) \n" + indent);
-        }
-    }
+        while (active) {
+            System.out.print("    You:");
+            String userInput = scanner.nextLine();
 
-    private void event(String[] parts, boolean isInit) throws LamballParseException {
-        // Checks if empty string (nothing after command) or only whitespaces
-        if (parts.length < 2 || parts[1] == null || parts[1].trim().isEmpty()) {
-            throw new LamballParseException(indent + "    Your event field is empty, baaaaka.\n" + indent);
-        }
-        String[] furtherSplit = parts[1].split(" /", 3);
-        if (furtherSplit.length < 3 || !furtherSplit[1].substring(0,5).equals("from ") ||
-                !furtherSplit[2].substring(0,3).equals("to ")) {
-            throw new LamballParseException(indent + "    Event is in the wrong formaaaaaaat, baa. :(\n    Correct " +
-                    "formaaat is: event <name> /from <time> /to <time>, baa.\n" + indent);
-        }
-        try {
-            Task temp = new Event(furtherSplit[0], furtherSplit[1].replaceFirst("from ", ""),
-                    furtherSplit[2].replaceFirst("to ", ""));
-            tasks.add(temp);
-            if (!isInit) {
-                System.out.println(indent + "    Added Event:\n        " + temp.toString() + "\n    Now you have "
-                        + tasks.size() + " tasks in the list.\n" + indent);
-                writeToFile("0 | " + temp.command());
-            }
-        } catch (DateTimeParseException e) {
-            throw new LamballParseException(indent + "    Dates are in the wrong formaaaaaaat, baa. :(\n    Correct fo" +
-                    "rmaaat is: yyyy-mm-dd (e.g 2001-01-20) \n" + indent);
-        }
-    }
-
-    private void deleteFromList(String[] parts) throws LamballParseException {
-        // If not numeric
-        if (!parts[1].matches("-?\\d+")) {
-            throw new LamballParseException(indent + "    Invalid number, baa.\n" + indent);
-        }
-        int idx = Integer.valueOf(parts[1]) - 1;
-        // Checks if index is within range of list
-        if (idx >= tasks.size() || idx < 0) {
-            throw new LamballParseException(indent + "    Taaask index out of range, baa.\n" + indent);
-        }
-        System.out.println(indent + "    I have removed this taaask:");
-        Task temp = tasks.remove(idx);
-        deleteLine(idx);
-        System.out.println("        " + temp.toString());
-        System.out.println("    Now you have " + tasks.size() + " tasks in the list.\n" + indent);
-    }
-
-    public boolean parse(String msg, boolean isInit) throws LamballParseException {
-        String[] parts = msg.split(" ", 2);
-        if (!isInit) {
-            System.out.println("\n    Lamball");
-        }
-        switch(parts[0]) {
-            case "mark": {
-                mark(parts, isInit);
-                return true;
-            }
-            case "unmark": {
-                unMark(parts);
-                return true;
-            }
-            case "bye":
-                goodbyeMessage();
-                return false;
-            case "list":
-                printList();
-                return true;
-            case "todo": {
-                toDo(parts, isInit);
-                return true;
-            }
-            case "event": {
-                event(parts, isInit);
-                return true;
-            }
-            case "deadline": {
-                deadline(parts, isInit);
-                return true;
-            }
-            case "delete": {
-                deleteFromList(parts);
-                return true;
-            }
-            default:
-                throw new LamballParseException(indent + "    Sorry, I don't understaaaaaand your commaaaaand, baa. :(\n" + indent);
-        }
-    }
-
-    public void obtainSavedFile() {
-        File folder = new File("src/main/java/data");
-        if (!folder.exists()) {
-            System.out.println("Folder does not exist. Creating folder...");
-            folder.mkdirs();
-        }
-        savedList = new File("src/main/java/data/list.txt");
-        try {
-            // Try to create file
-            if (savedList.createNewFile()) {
-                System.out.println("List created successfully at: " + savedList.getAbsolutePath());
-            } else {
-                System.out.println("Seems like I haave helped you before, so no new list is needed!");
-            }
-            this.initializeListFromText();
-        } catch (IOException e) {
-            System.err.println("An error occurred while creating the file: " + e.getMessage());
-        }
-    }
-
-    private static void writeToFile(String filePath, String toAdd) {
-        try {
-            FileWriter fw = new FileWriter(filePath, true);
-            fw.write(toAdd + System.lineSeparator());
-            fw.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-            System.out.println("Save Failed: " + e.getMessage());
-        }
-    }
-    private static void writeToFile(String toAdd) {
-        String filePath = "src/main/java/data/list.txt";
-        try {
-            FileWriter fw = new FileWriter(filePath, true);
-            fw.write(toAdd + System.lineSeparator());
-            fw.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-            System.out.println("Save Failed: " + e.getMessage());
-        }
-    }
-
-    private void initializeListFromText() throws FileNotFoundException {
-        File tempFile = new File("src/main/java/data/tempfile.txt");
-        try {
-            // Delete and creates a new tempfile
-            tempFile.delete();
-            tempFile.createNewFile();
-        } catch (IOException e) {
-            System.out.println(e.getMessage());
-        }
-        Scanner scanner = new Scanner(savedList);
-        System.out.println("Initializing saved file...");
-        int count = 0;
-        while(scanner.hasNext()) {
+            // Echo the user's command
             try {
-                // Parses every line of the saved file - if error, deletes line in the file and ignores command
-                String currLine = scanner.nextLine();
-                String[] parts = currLine.split(" \\| ", 2);
-                // Means that it is not formatted correctly in the <0 or 1> | <command> format
-                if (parts.length != 2 || !(Integer.valueOf(parts[0]) == 1 || Integer.valueOf(parts[0]) == 0)) {
-                    throw new LamballParseException("Corrupt format, ignoring...");
+                active = lamball.parse(userInput, false);
+                if (!active) {
+                    Ui.goodbyeMessage();
                 }
-                this.parse(parts[1], true);
-                count++;
-                // Marks task if first character is 1. Else does not.
-                if (Integer.valueOf(parts[0]) == 1) {
-                    this.parse("mark " + count, true);
-                }
-                // If code reaches here, means that the line is valid - write to temp file
-                writeToFile("src/main/java/data/tempfile.txt", currLine);
-
-
             } catch (LamballParseException e) {
-                // Ignores line
-                System.out.println("Corrupt format, ignoring...");
+                System.out.println(e.getMessage());
             }
+
         }
         scanner.close();
-        System.out.println("Done!");
-        savedList.delete();
-        tempFile.renameTo(savedList);
     }
-
-    public static void replaceLine(String toWrite, int index) {
-        String filePath = "src/main/java/data/list.txt";
-//        System.out.println(toWrite);
-//        System.out.println(index);
-        try {
-            // Read all lines from the file
-            Path path = Paths.get(filePath);
-            List<String> lines = Files.readAllLines(path);
-
-            // Check if the index is valid
-            if (index >= 0 && index < lines.size()) {
-                // Replace the line at the specified index
-                lines.set(index, toWrite);
-
-                // Write the modified content back to the file
-                Files.write(path, lines);
-
-//                System.out.println("Line " + index + " replaced successfully.");
-            } else {
-//                System.err.println("aaaaaaa");
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) {
+        new Lamball().run();
     }
-    public static void deleteLine(int index) {
-        String filePath = "src/main/java/data/list.txt";
-        try {
-            // Read all lines from the file
-            Path path = Paths.get(filePath);
-            List<String> lines = Files.readAllLines(path);
-
-            // Check if the index is valid
-            if (index >= 0 && index < lines.size()) {
-                // Remove the line at the specified index
-                lines.remove(index);
-
-                // Write the modified content back to the file
-                Files.write(path, lines);
-
-//                System.out.println("Line " + index + " deleted successfully.");
-            } else {
-                System.err.println("Invalid line number.");
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-
 }
+
+
